completed 4 sprint

+++ src/sessionLogic/securityDevices/securityDevices.repository.ts
@@ -16,39 +16,66 @@ export class SecurityDevicesRepository{
     }
 
     async findSessionByDeviceId(deviceId:string){
-        const session= await this.sessionModel.findOne({deviceId: deviceId});
-        return session;
+        const session= await this.sessionModel.findOne({deviceId: deviceId}).lean<Session>();
+        return session
+    }
+    async findSessionByDeviceIdAndUserId(deviceId:string, userId: string){
+        const session= await this.sessionModel.findOne({deviceId: deviceId, userId: userId}).lean<Session>();
+        return session
     }
 
     async findSessionForRefresh(userId: string, deviceId:string){
-        const session= await this.sessionModel.findOne({userId: userId,deviceId: deviceId, revoked: false}).lean();
+        const session= await this.sessionModel.findOne({userId: userId,deviceId: deviceId}).lean<Session>();
         return session;
     }
     async findFrontSessionByDeviceId(deviceId:string){
+        console.log(
+            await this.sessionModel.find().lean()
+        );
         const session= await this.sessionModel.findOne({deviceId: deviceId}).lean<Session>();
         return session;
     }
+    async findSession(userId: string, deviceId:string, sessionVersion: number){
+        const session= await this.sessionModel.findOne({userId: userId,deviceId: deviceId, version: sessionVersion}).lean<Session>();
+        return session;
+    }
+    async findSessionForLogout(userId: string, deviceId:string){
+        const session= await this.sessionModel.findOne({userId: userId,deviceId: deviceId}).lean<Session>();
+        return session;
+    }
 
-    async updateSession(deviceId:string, lastActivity:Date, expiresAt:Date) {
+    async updateSession(deviceId:string, userId: string, lastActivity:Date, expiresAt:Date, sessionVersion: number) {
         const result = await this.sessionModel.updateOne(
-            { deviceId: deviceId },
+            { userId: userId, deviceId: deviceId },
             {
                 $set: {
                     lastActivity: lastActivity,
                     expiresAt: expiresAt,
+                    version: sessionVersion
                 },
             },
         );
 
-        return result.matchedCount === 1 && result.modifiedCount === 1;
+        return result.acknowledged && result.matchedCount === 1;
+    }
+    async updateRevokedSession(userId: string, deviceId: string, lastActivity:Date, sessionVersion: number) {
+        const result = await this.sessionModel.updateOne(
+            { userId: userId, deviceId: deviceId, version: sessionVersion},
+            {
+                $set: {
+                    lastActivity: lastActivity,
+                    revoked: true
+                },
+            },
+        );
+        return result.acknowledged && result.matchedCount === 1;
     }
 
-    async closeSession(deviceId: string){
-        console.log('DELETE session deviceId:', deviceId);
-        const result = await this.sessionModel.deleteOne({deviceId: deviceId });
-        console.log('DELETE result:', result);
 
-        return result.deletedCount === 1
+    async closeSession(userId: string, deviceId: string){
+        const result = await this.sessionModel.deleteOne({userId: userId, deviceId: deviceId });
+        //return result.deletedCount === 1
+        return
     }
 
     async closeAllSessionsBesidesThisOne(userId:string, deviceId:string){
